Updates to Layout component, including shared MainContent and Sidebar components
