+++ src/components/common/Layout/index.js
@@ -1,0 +1,3 @@
+export { default as Layout } from './Layout';
+export { default as MainContent } from './MainContent';
+export { default as Sidebar } from './Sidebar';
