+++ src/components/common/with-authentication/withAuthentication.jsx
