+++ src/components/masters/programs-list/ProgramListPage.jsx
@@ -1,13 +1,13 @@
 import React, { Component } from 'react';
 import { navigate } from 'gatsby';
 import PropTypes from 'prop-types';
-import { IntlProvider } from 'react-intl';
 import { connect } from 'react-redux';
 import { StatusAlert } from '@edx/paragon';
 import { faExclamationTriangle } from '@fortawesome/free-solid-svg-icons';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 
-import { Layout, withAuthentication } from '../../common';
+import { withAuthentication } from '../../common/with-authentication';
+import { Layout } from '../../common/layout';
 
 import { fetchUserProgramEnrollments } from '../user-program-enrollments';
 
@@ -88,7 +88,7 @@ export class ProgramListPage extends Component {
     const { validPrograms } = this.state;
 
     return (
-      <IntlProvider locale="en">
+      <>
         {isLoading ? (
           <Layout>
             <div className="container my-4">
@@ -155,7 +155,7 @@ export class ProgramListPage extends Component {
             )}
           </>
         )}
-      </IntlProvider>
+      </>
     );
   }
 }
